// src/app/enums/enums.ts
 
  export enum AngularAlertType {
    Swal = 1,
    Toast = 2
  }

  export enum UserType
  {
      Candidate = 1,
      PartySecretaray = 2,
      ElectionOfficer = 3
  }
  export enum InvitationStatus
  {
      ReadyToSend = 1,
      Sent = 2,
      Canceled = 3,
      Confirmed = 4,
      Rejected = 5,
      Deleted = 6
  }