// src/app/models/partyList/Invitation/invitation.ts
export class invitation {
  candidateId: string;
  invitationId: number;
  publicationListId: number;
  code: number;
  firstName: string;
  lastName: string;
  fatherName: string;
  cellPhoneNumber: string;
  ranking: number;
  status: number;
  statusTitle: string;
  description: string;
  invitationDate: string;
  partyNam: string;
}